Separate nodes out to separate submodules
